+++ src/Factory.php
@@ -24,33 +24,60 @@ interface Factory
     public function getDefaultShortFormat(): Format;
 
     public function createVoid(): Uuid;
+    public function createVoidString(): string;
 
     public function createV1(
         int|string|null $node = null,
         ?int $clockSeq = null
     ): Uuid;
 
+    public function createV1String(
+        int|string|null $node = null,
+        ?int $clockSeq = null
+    ): string;
+
     public function createV3(
         string $name,
         ?string $namespace = null
     ): Uuid;
 
+    public function createV3String(
+        string $name,
+        ?string $namespace = null
+    ): string;
+
     public function createV4(): Uuid;
+    public function createV4String(): string;
     public function createV4Comb(): Uuid;
+    public function createV4CombString(): string;
 
     public function createV5(
         string $name,
         ?string $namespace = null
     ): Uuid;
 
+    public function createV5String(
+        string $name,
+        ?string $namespace = null
+    ): string;
+
     public function createV6(
         int|string|null $node = null,
         ?int $clockSeq = null
     ): Uuid;
 
+    public function createV6String(
+        int|string|null $node = null,
+        ?int $clockSeq = null
+    ): string;
+
     public function createV7(
         ?DateTimeInterface $date = null
     ): Uuid;
+
+    public function createV7String(
+        ?DateTimeInterface $date = null
+    ): string;
 
 
     public function isValid(